sorted users and classes function for different files and removed unnecessory variables declared in users.js which will be using unncessaory spaces
