+++ package.json
@@ -3,7 +3,8 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "node ./bin/www"
+    "start": "nodemon ./bin/www",
+    "dev": "nodemon ./app.js localhost 3000"
   },
   "dependencies": {
     "cookie-parser": "~1.4.4",
@@ -11,7 +12,9 @@
     "express": "~4.16.1",
     "http-errors": "~1.6.3",
     "jade": "~1.11.0",
-    "morgan": "~1.9.1",
+    "morgan": "~1.9.1"
+  },
+  "devDependencies": {
     "nodemon": "^2.0.4"
   }
 }
